refactor: 패키지 변경 - 모델 클래스들을 도메인 패키지로 이동
