+++ src/main/java/lotto/domain/Lotto.java
@@ -1,12 +1,8 @@
-package lotto;
+package lotto.domain;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class Lotto {
