+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,6 +1,6 @@
-package lotto;
+package lotto.domain;
 
-import static lotto.LottoFixture.*;
+import static lotto.domain.LottoFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
